git-svn-id: https://bsisvn.mayo.edu/main/projects/g/Genome_GPS/trunk@577 8f329c14-d232-4f17-8684-7ee34322b8dc

+++ merge_raw_variants.sh
@@ -14,6 +14,10 @@ else
 	ref=$( cat $tool_info | grep -w '^REF_GENOME' | cut -d '=' -f2)
 	gatk=$( cat $tool_info | grep -w '^GATK' | cut -d '=' -f2)
 	java=$( cat $tool_info | grep -w '^JAVA' | cut -d '=' -f2)
+	vcftools=$( cat $tool_info | grep -w '^VCFTOOLS' | cut -d '=' -f2)
+	perllib=$( cat $tool_info | grep -w '^PERLLIB_VCF' | cut -d '=' -f2)
+	tabix=$( cat $tool_info | grep -w '^TABIX' | cut -d '=' -f2)
+	PERL5LIB=$perllib
 	
 	var_dir=$output_dir/variants
 	inputargs=""
@@ -23,11 +27,11 @@ else
 		input=$var_dir/$sample
 		for chr in $chrs
 		do
-			inputfile=$input/$sample.variants.chr$chr.raw.all.vcf
-			input_indexfile=$input/$sample.variants.chr$chr.raw.all.vcf.idx
+			inputfile=$input/$sample.variants.chr$chr.raw.all.vcf.gz
+			input_indexfile=$input/$sample.variants.chr$chr.raw.all.vcf.gz.idx
 			if [ -s $inputfile ]
 			then
-				inputargs="-V $inputfile "$inputargs
+				inputargs="$inputfile "$inputargs
 				input_index=" $input_indexfile"$input_index 
 			else
 				echo "WARNING : $inputfile not there"
@@ -36,25 +40,31 @@ else
 	done
 
 	### merge all the variants using GATK
-	$java/java -Xmx2g -Xms512m -jar $gatk/GenomeAnalysisTK.jar \
-	-R $ref \
-	-et NO_ET \
-	-T CombineVariants \
-	$inputargs \
-	-o $var_dir/raw.vcf
-	### raw SNV
-	cat $var_dir/raw.vcf | awk '(length($4) == 1 && length($5) == 1 ) || $0 ~ /#/' > $var_dir/raw.SNV.vcf  
-	### raw INDEL
-	cat $var_dir/raw.vcf | awk '(length($4) > 1 || length($5) > 1 ) || $0 ~ /#/' > $var_dir/raw.INDEL.vcf 
+#	$java/java -Xmx2g -Xms512m -jar $gatk/GenomeAnalysisTK.jar \
+#	-R $ref \
+#	-et NO_ET \
+#	-T CombineVariants \
+#	$inputargs \
+#	-o $var_dir/raw.vcf
 	
-	rm $var_dir/raw.vcf
-	gzip $var_dir/raw.SNV.vcf   
-	gzip $var_dir/raw.INDEL.vcf 
-	#gzip $var_dir/raw.SNV.vcf
+	$vcftools/bin/vcf-merge $inputargs > $var_dir/raw.vcf.gz 
+	
+
+	### raw SNV
+	$tabix/bgzip -c -d $var_dir/raw.vcf.gz | awk '(length($4) == 1 && length($5) == 1 ) || $0 ~ /#/' > $var_dir/raw.SNV.vcf  
+	### raw INDEL
+	cat $var_dir/raw.vcf.gz | awk '(length($4) > 1 || length($5) > 1 ) || $0 ~ /#/' > $var_dir/raw.INDEL.vcf 
+	
+	if [ -s $var_dir/raw.SNV.vcf && -s $var_dir/raw.INDEL.vcf ]
+    then
+        rm $var_dir/raw.vcf
+        $tabix/bgzip $var_dir/raw.SNV.vcf   
+        $tabix/bgzip $var_dir/raw.INDEL.vcf 
+	fi
+
 	if [ -s $var_dir/raw.SNV.vcf.gz ]
 	then
-		file=`echo $inputargs | sed -e '/-V/s///g'`
-		rm $file
+		rm $inputargs
 		rm $input_index
 	fi	
 	echo `date`
